backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Pinned LangChain ecosystem (mutually compatible)
langchain==0.2.12
langchain-community==0.2.11
langchain-openai==0.1.21
langchain-text-splitters==0.2.2

# Model SDKs
openai==1.42.0

# PDF + Vector store
pypdf==3.17.0
faiss-cpu==1.8.0.post1

# Infra / misc
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.4
websockets==12.0
aiofiles==23.2.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
google-search-results==2.4.2
requests==2.31.0
httpx==0.27.0